+++ src/Console/SolanaStaking/Main.hs
@@ -10,13 +10,15 @@ module Console.SolanaStaking.Main
     ) where
 
 import           Control.Monad                  ( (<=<)
+                                                , foldM
                                                 , forM
                                                 , unless
                                                 )
 import           Control.Monad.Except           ( ExceptT
                                                 , runExceptT
                                                 )
-import           Control.Monad.Reader           ( ReaderT
+import           Control.Monad.Reader           ( MonadIO
+                                                , ReaderT
                                                 , liftIO
                                                 , runReaderT
                                                 )
@@ -24,7 +26,14 @@ import           Data.Bifunctor                 ( bimap
                                                 , second
                                                 )
 import           Data.List                      ( sortOn )
-import           Data.Maybe                     ( fromMaybe )
+import           Data.Maybe                     ( fromMaybe
+                                                , mapMaybe
+                                                )
+import           Data.Time.Clock.POSIX          ( posixSecondsToUTCTime )
+import           Data.Time                      ( LocalTime (..)
+                                                , ZonedTime (..)
+                                                , utcToLocalZonedTime
+                                                )
 import           Data.Version                   ( showVersion )
 import           System.Console.CmdArgs         ( (&=)
                                                 , Data
@@ -51,6 +60,8 @@ import           Console.SolanaStaking.Csv      ( makeCsvContents )
 import           Paths_solana_staking_csvs      ( version )
 
 import qualified Data.ByteString.Lazy          as LBS
+import qualified Data.Text                     as T
+import qualified Data.Map.Strict               as M
 
 
 -- | Pull staking rewards data for the account & print a CSV to stdout.
@@ -63,7 +74,6 @@ run Args {..} = either (error . show) return <=< runner $ do
     -- Grab rewards for each stake account
     (stakeErrors, stakeRewards) <-
         fmap (bimap concat concat . unzip) . forM stakes $ \sa -> do
-
             second (map (sa, ))
                 <$> maybe (getAllStakeRewards (saPubKey sa))
                           (getYearsStakeRewards $ saPubKey sa)
@@ -73,17 +83,48 @@ run Args {..} = either (error . show) return <=< runner $ do
         mapM_ (hPutStrLn stderr . ("\t" <>) . show) stakeErrors
     -- Write the CSV
     let orderedRewards = sortOn (srTimestamp . snd) stakeRewards
+        aggregator = if argAggregate then aggregateRewards else return
         outputFile     = fromMaybe "-" argOutputFile
+    rewards <- aggregator orderedRewards
     if argCoinTracking
-        then liftIO $ makeCoinTrackingImport outputFile orderedRewards
+        then liftIO $ makeCoinTrackingImport outputFile rewards
         else do
-            let output = makeCsvContents orderedRewards
+            let output = makeCsvContents rewards
             if outputFile == "-"
                 then liftIO $ LBS.putStr output
                 else liftIO $ LBS.writeFile outputFile output
   where
     runner :: ReaderT Config (ExceptT APIError IO) a -> IO (Either APIError a)
     runner = runExceptT . flip runReaderT (mkConfig argApiKey argPubKey)
+    aggregateRewards :: MonadIO m => [(StakingAccount, StakeReward)] -> m [(StakingAccount, StakeReward)]
+    aggregateRewards =
+        fmap (mapMaybe (aggregate . snd) . M.toList)
+            . foldM (\acc r -> do
+                        rewardTime <- liftIO . utcToLocalZonedTime $ posixSecondsToUTCTime $ srTimestamp $ snd r
+                        return $ M.insertWith
+                            (<>)
+                            (localDay $ zonedTimeToLocalTime rewardTime)
+                            [r]
+                            acc
+                    )
+                    M.empty
+
+    aggregate :: [(StakingAccount, StakeReward)] -> Maybe (StakingAccount, StakeReward)
+    aggregate = \case
+        [] -> Nothing
+        rs -> Just
+            ( StakingAccount
+                { saValidatorName = "AGGREGATE-" <> T.pack (show $ length rs)
+                , saPubKey = StakingPubKey $ T.pack argPubKey
+                , saLamports = sum $ map (saLamports . fst) rs
+                }
+            , StakeReward
+                { srTimestamp = minimum $ map (srTimestamp . snd) rs
+                , srSlot= srSlot $ snd $ head rs
+                , srEpoch= srEpoch $ snd $ head rs
+                , srAmount= sum $ map (srAmount . snd) rs
+                }
+            )
 
 
 -- | CLI arguments supported by the executable.
@@ -100,6 +141,8 @@ data Args = Args
     -- Imports.
     , argYear         :: Maybe Integer
     -- ^ Year to limit output to.
+    , argAggregate    :: Bool
+    -- ^ Aggregate rewards into single-day transactions.
     }
     deriving (Show, Read, Eq, Data, Typeable)
 
@@ -129,6 +172,10 @@ argSpec =
                                 &= name "y"
                                 &= name "year"
                                 &= typ "YYYY"
+            , argAggregate    = False
+                                &= help "Output one aggregate row per day."
+                                &= explicit
+                                &= name "aggregate"
             }
         &= summary
                (  "solana-staking-csvs v"