+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Add `--aggregate` flag to group rewards by day.
 * Fetch rewards until no more rewards are returned instead of stopping when
   less than the 5-item page limit is returned.
 * Handle breaking API changes when rate limited.